requirements.txt
pygame==1.9.3
click==6.7
vtjp==0.1.14
currencyconverter==0.13.3
python-dotenv==0.8.2
google-api-python-client
python-dateutil
